Move metrics module into sim package
